requirements.txt
# requirements.txt
# Foundational dependencies for Mario RL agent project

# requirements.txt
# Foundational dependencies for Mario RL agent project

numpy==1.24.4
gym-super-mario-bros
gymnasium
setuptools
wheel
